+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # CHANGELOG
 
-## UNRELEASED
+## 0.3.0 (UNRELEASED)
 
 - Added `CacheSerializer`, `NoopCacheSerializer` and `JSONCacheSerializer`. Changed `CacheBackend`, `InMemoryCache`, `CloudflareCacheBackend` and `DynamoDBCacheBackend` to accept `serializer` initialization option.
 - Fixed schema proxy returning an error when variable defined in an operation is missing from its variables.
 - Improved custom headers handling in `ProxyResolver` and `ProxySchema`.
+- Added fields dependencies configuration option to `ProxySchema`.
 
 
 ## 0.2.0 (2023-09-25)